restore op_kernel since it will be moved to op_kernel_type
